pyproject.toml
[project]
name = "anemoi-weathermart-plugins"
version = "0.1.0"
description = "Package connecting custom sources to anemoi-datasets interface."
readme = "README.md"
authors = [{ name = "Ophélia Miralles" }]
requires-python = ">=3.11"
dependencies = [
    "anemoi-datasets",
    "anemoi-plugins",
    "anemoi-transform",
    "pyproj",
    "pandas",
    "xarray",
    "zarr",
    "dask",
    "pyyaml",
    "omegaconf>=2.3.0,<3",
    "weathermart",
]


[project.scripts]
anemoi-weathermart-plugins = "anemoi_weathermart_plugins:main"

[project.entry-points."anemoi.datasets.sources"]
satellite = "anemoi_weathermart_plugins.transform.sources:SATELLITE"
dem = "anemoi_weathermart_plugins.transform.sources:NASADEM"
opera = "anemoi_weathermart_plugins.transform.sources:OPERA"

[project.entry-points."anemoi.transform.filters"]
interp2grid = "anemoi_weathermart_plugins.transform.filters:Interp2Grid"
interp2res = "anemoi_weathermart_plugins.transform.filters:Interp2Res"
interpNA = "anemoi_weathermart_plugins.transform.filters:InterpNAFilter"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dependency-groups.dev]
dependencies = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "mypy>=1.6.1,<2",
    "pydata-sphinx-theme>=0.14.1,<0.15",
    "pylint>=3.0.2,<4",
    "pytest-cov>=4.1.0,<5",
    "sphinx>=7.2.6,<8",
    "sphinx-autoapi>=3.0.0,<4",
    "yapf>=0.40.2,<0.41",
]

[tool.uv.sources]
anemoi-plugins = { git = "https://github.com/ecmwf/anemoi-plugins.git" }
anemoi-datasets = { git = "https://github.com/ecmwf/anemoi-datasets.git"}
anemoi-transform = { git = "https://github.com/ecmwf/anemoi-transform.git" }
weathermart = {git = "https://github.com/OpheliaMiralles/weathermart.git"}


[[tool.uv.index]]
name = "pypi-mch"
url = "https://service.meteoswiss.ch/nexus/repository/python-all/simple"

[[tool.uv.index]]
name = "pypi-mch-publish"
url = "https://service.meteoswiss.ch/nexus/repository/python-mch/"

[tool.yapf]
based_on_style = "pep8"
column_limit = "120"

[tool.pylint.master]
disable = [
    'R0913', # too many arguments
    'R0917', # too many positional arguments
    'C0415', # import outside toplevel
    'C0414', # explicit reexport
    'C0114', # missing-module-docstring
    'C0115', # missing-class-docstring
    'C0116', # missing-function-docstring
    'W0511', # fix me or to-do comments are already covered by SonarQube
    'C0414', # useless-import-alias
]

[tool.pylint.format]
# Maximum number of lines in a module.
max-line-length = 120

[tool.pylint.design]
# Minimum number of public methods for a class (see R0903).
min-public-methods = 0

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.coverage.run]
omit = ["test/**", "doc/**"]
